requirements_deploy.txt
fabric
requests
GitPython==0.3.2.RC1
